+++ src/graphql/pubsub/index.js
@@ -1,5 +1,6 @@
 const { PubSub } = require('apollo-server-express')
 const bindUserEvents = require('./user')
+const bindFirmwareEvents = require('./firmware')
 const bindMeasurementEvents = require('./measurement')
 
 const pubsub = new PubSub()
@@ -8,6 +9,7 @@ if (process.env.NODE_ENV !== 'test') {
 	// which does not support change
 	// streams, so skip if testing
 	bindUserEvents(pubsub)
+	bindFirmwareEvents(pubsub)
 	bindMeasurementEvents(pubsub)
 }